구간 등록 및 삭제에 대한 책임을 Sections가 갖게하라
