+++ src/main/java/subway/domain/Line.java
@@ -60,26 +60,7 @@ public class Line {
         this.sections.remove(section);
     }
 
-    public void verifyAddableSection(Section section) {
-        if(!section.getUpStation().equals(sections.getLastSection().getDownStation())) {
-            throw new SectionAddFailureException("새로운 구간의 상행역이 기존 구간의 하행역이 아닙니다.");
-        }
 
-        if (sections.isAlreadyExistStation(section.getDownStation())) {
-            throw new SectionAddFailureException("새로운 구간의 하행역이 기존 노선에 이미 존재합니다.");
-        }
-    }
-
-    public void verifyDeletableStation(Station station) {
-        if (sections.hasOnlyOneSection()) {
-            throw new SectionDeleteFailureException("노선의 구간은 최소 한 개 이상 존재해야 합니다.");
-        }
-
-        Section lastSection = sections.getLastSection();
-        if (lastSection == null || !lastSection.getDownStation().equals(station)) {
-            throw new SectionDeleteFailureException("노선의 하행종점역만 제거할 수 있습니다.");
-        }
-    }
 
     public List<Station> getStations() {
         return this.sections.getStations();
